Add Employees endpoints
So that we can correctly assign sickness in 10kft we need to be able
to query the employee endpoint in Breathe to get the email address
associated.

The Breathe client was updated to allow for fetching a single resource
(ie. with no arguments and with an id appended to the path).

It's worth noting that Breathe returns an array of employees even when
fetching by ID so I've reused the `employees.json` fixture.